--- pages/products/[id].vue
@@ -1,9 +1,0 @@
-<template>
-  <div>
-    {{ id }}
-  </div>
-</template>
-
-<script setup>
-const { id } = useRoute().params
-</script>
